Change FBAOrder.box_weight to DecimalField

+++ fba/models.py
@@ -108,7 +108,9 @@ class FBAOrder(models.Model):
     box_width = models.PositiveIntegerField(blank=True, null=True)
     box_height = models.PositiveIntegerField(blank=True, null=True)
     box_depth = models.PositiveIntegerField(blank=True, null=True)
-    box_weight = models.PositiveIntegerField(blank=True, null=True)
+    box_weight = models.DecimalField(
+        max_digits=5, decimal_places=2, blank=True, null=True
+    )
     notes = models.TextField(blank=True)
     priority = models.PositiveIntegerField(default=999)
     printed = models.BooleanField(default=False)
